+++ sidebars.js
@@ -409,6 +409,20 @@ const sidebars = {
                 "common-workflows/analytics/manage-insights/update-insight-status",
               ],
             },
+            {
+              type: "category",
+              label: "Use Bot APIs",
+              items: [
+                "common-workflows/analytics/use-bot-api/use-bot-api",
+                "common-workflows/analytics/use-bot-api/get-bot-configuration",
+                "common-workflows/analytics/use-bot-api/get-bot-question-list",
+                "common-workflows/analytics/use-bot-api/create-a-bot-instance",
+                "common-workflows/analytics/use-bot-api/ask-bot-instance-suggested-question",
+                "common-workflows/analytics/use-bot-api/ask-bot-instance-question",
+                "common-workflows/analytics/use-bot-api/get-a-bot-question",
+                "common-workflows/analytics/use-bot-api/delete-a-bot-instance",
+              ],
+            },
           ],
         },
         {